Update patch set 2
Patch Set 2:

I don't know where it is shown to be honest. I don't have any Miracast-enabled devices to test on. I happened to stumble upon the string when making the other changes. It could be that it was never used or that it might have been updated later. This change is from January som things could have changed since then. As long as "display" is not translated "visas" I'm ok with whatever change you make.

Patch-set: 2