Update patch set 1
Patch Set 1: Verified-1

The underlying code was refactored away in Google's long-term development branch. We won't be able to merge this change as-is. Reviewers can investigate whether this might need to be manually re-implemented in the refactored implementation. There won't be any chance to successfully submit this change through AOSP in a way that merges with Google's internal tree.

Patch-set: 1
Reviewer: Gerrit User 1000660 <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=-1