Update patch set 1
Patch Set 1: Code-Review-2

This change violates at least half a dozen invariants of the power manager.  It also significantly affects user experience.

The stated behavior is that the system does not go to sleep until the display timeout occurs while prox is active.  We could shorten the display timeout in this case but we must not just go to sleep immediately.  It will be disruptive to the user.

It is also very important that we continue to hold a wake lock while prox sensor is in use.  We can't change that behavior unless we improve the sensor HAL to treat prox as a wakeup sensor.

Patch-set: 1
Reviewer: Gerrit User 1003957 <1003957@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2