DatePicker: only trigger OnDateSetListener if the date has actually changed
Previously, the onDateSetListener was triggered when someone clicked on
the positive action button, but it was also triggered in the onStop()
when the dialog was closed. So even if someone cancelled the dialog, or added
their own negative button, the onDateSetListener was triggered incorrectly.
And if someone had clicked on the positive button, the listener was
triggered twice. Fixed to no longer trigger during the onStop() event.

Bug: https://code.google.com/p/android/issues/detail?id=34833
Change-Id: I4efa4d2abda4ad784a6a5189b823fdcacfad9347
Signed-off-by: howardb <android@howardb.com>