Update patch set 1
Patch Set 1:

The call to the OnDateSetListener means that a date was selected via the UI not that the date was confirmed. From the JavaDoc of the OnDateSetListener "The callback used to indicate the user is done filling in the date." does not imply the user has confirmed the date. This is why the listener is named OnDateSetListener no OnDateConfirmedListener (or similar).

It is responsibility of the client to discard the date if the dialog was cancelled. This is why dialog has API to register an OnCancelListener. Also changing behavior at this point is risky since apps may be relying on that. Last but not least, the UX team requested this behavior.

I agree that if the date did not change calling back is not necessary.

Patch-set: 1