Update patch set 1
Patch Set 1:

so in the context of Settings, are you saying that the UI folks consider it not a bug that hitting the back button in the "set date" dialog doesn't cancel? or that we need to add extra code to Settings to "undo" if the user hits back?

Patch-set: 1