Update patch set 1
Patch Set 1:

I stumbled over this when I recently wanted to monitor an app that downloads a number of files and throws a new notification for each one: It was outright impossible to open the notification area at all. As the user can't control the timing of notification clearing by the user, IMHO the auto-close feature should be strictly limited to swiping away the notifications.

Patch-set: 1
Reviewer: Gerrit User 1010292 <1010292@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0