Add system partition check for vpn dialogs
The VPN permission check required internal knowledge that
other checks in the system would ensure that
package "com.android.vpndialogs" was genuine. In the
off chance those other checks change, or someone is
able to spoof the package name, this will at least
check to see that the app is on the system image; one
more hurdle to jump.

Signed-off-by: William Roberts <wroberts@tresys.com>

Change-Id: I4875a4a30bea8975943c889d3b1ebefc84c07ee9