Add a new getAllCheckedItemPositions() method to get all checked positions
The API already includes a getCheckedItemPositions() method but it makes
the API unclear as the return type is rather opaque to Android
developers. Indeed, it returns a set a key-value pairs but the set of 
all checked items is made of all keys having a value of true.

This addition modifies the API but returns a clear and understandable
array of int containing all checked positions.

Note: I've never committed a patch modifying the API. I guess, this the
process is a little bit different: method naming validation, 
documentation, etc. Please feel free to tell me what to do.

I'm personally not satisfied with the naming but couldn't find a better
name than the already used getCheckedItemPositions() which could be 
deprecated to my mind. (not added in this patch)

Change-Id: I2100258e1682e3896fdf14436fad459f5d72dc82