wifi: Fix for coonection between GO and p2p device.
When GO sends an invitation request to the p2p device without
persistent profile, GO gets an invitation result=8 which means
and unknown group, when GO receives invitation result 8, its tries
to remove persistent profile of the client but it will not be able
find the profile so it does continue with sending the invitation
so the connection fails. To address this, ignore if there are no
persistent profiles found for the client and continue with sending
the Invitation request to the client to make connection succesful
connection.

Change-Id: I78f54d4bb2f0d603bc2170ec2c3ef55a91050236