Update patch set 2
Patch Set 2:

I didn't look in detail, because this is a big change. At a glance, the overall code style is fine, but there are some trailing whitespaces that should be removed, and some lines that should be wrapped to 100 characters.

The change merges and builds fine in Google's internal master tree. Untested.

Patch-set: 2
Reviewer: Gerrit User 1000660 <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0