Update patch set 3
Patch Set 3: Code-Review-1

Santos,

Here is Antonio suggestion to expose precise information about the call state, see how this fits into the refactoring of the phone app. What we should strive to do is maximize the phone app's use of TelephonyManager and minimize its use of telephony internals. With the use of internals eventually going to zero, I see this as a step that right direction.

More comments to come but -1, while we review.

Patch-set: 3
Label: Code-Review=-1