Update patch set 1
Patch Set 1: Code-Review-2

Whether or not it is okay to change the API, this is not the proper way to do it -- the only api definition that should be changed is the current one, not everything back to the end of time.

Furthermore, we don't really have a way to correctly accept API changes in AOSP.  The process for this is that we finalize the most recently released API, and that is N.txt.  Then we start on the next version, copying N.txt to N+1.txt.  Any changes to APIs go in to N+1.txt, since that is the platform version where they are changing.

You can not change previous version definitions like this.  Doing so causes the API diffs to be wrong etc.

Patch-set: 1
Reviewer: Gerrit User 1000559 <1000559@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2