Update patch set 1
Patch Set 1:

Thanks for you comments Dianne.

Making an incompatible change to an API is a hassle and it always bites irrespective of the approach. The main reasons to rewrite the history was that it's bad to just break the api from version N to N+1 (though it's also bad to pretend...) and that this change might not break any existing app actually, or a very small number. We were trying to pick the lesser evil...

If there's no correct way to accept api changes through AOSP could you share what would be the right steps? The other changes are submitted though the AOSP.

Patch-set: 1
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0