Update patch set 1
Patch Set 1:

No, I downloaded the patch in the link and the problem we detected is still present.

The problem is a bout an exception thrown when right clicking with a mouse outside the images in a grid.

To reproduce, connect mouse via usb, open for example the vanilla browser, go into bookmark menu, right click on an area outside the images and the browser will force stop.

Patch-set: 1
Reviewer: Gerrit User 1004424 <1004424@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0