Update patch set 1
Patch Set 1:

Only a subset of JNI functions are allowed between critical "get" and "release".  This is only enforced when CheckJNI is enabled.  The idea is that, if you really are set on obstructing the garbage collector, you shouldn't be calling NewStringUTF or CallVoidMethod while doing so.

In the current implementation of Dalvik, the critical and non-critical implementations are effectively identical.  There is no guarantee that this will hold in the future.

Patch-set: 1
Reviewer: Gerrit User 1000483 <1000483@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0