Remove opaque check in preparing dirty region
Since preserve swap is enabled, we need to clear the color buffer
when the scissor rect is prepared for a new process. This prevents
garbage being present from the previous process in the color buffer.

issue:
without this visual corruption seen in google maps usecase
create short-cut on home screen and open maps
use back button to go to home screen
click gmap to open again to see the 1-2 frame corruption

this can be very easily seen on nexus-4
and with some efforts on nexus-5 and nexus-6
in gltrace it would be easy to spot the corruption.

Change-Id: Icd12ae388077b8c9ed329c37314e896500078543