Assure StorageNotification passes correct volume for formatting
If a format request comes to the ExternalStorageFormatter, it
will check the intent for the extra 'volume' data. If not present,
it will use the default external storage path (/storage/emulated/legacy).
This causes an IllegalArgumentException and a soft reboot. The change
assures that StorageNotification also sends in the intent the volume to
be formatted (eg./storage/usbcard) for which the notification was
received in the first place.

Change-Id: I25e0a367726cc20fd5d0aed8e129d587298f3e0f
Signed-off-by: Adriana Reus <adriana.reus@intel.com>