Print dispatch history of mainLooper to ease ANR debugging
Dispatch history of mainLooper is important for debugging.
With these logs, ANR issues can be triaged more efficiently.
With these logs, tool can be developed to debug ANR issues.

First, ANR may account for about 10% of the bugs reported
during device development. It is important to improve the
debugging efficiency for these bugs. With these logs, ANR
analyzer can be developed to debug ANR issues automatically.
With advanced debugging skills programmed into the analyzer,
text report and visual profiling graph can be produced to
help identify the culprit more accurately, easier, faster.
Second, huge number of developers are involved in Android
development. Many of them are not skilled at debugging all
sorts of ANR issues. These logs and tools definitely would
reduce the complexity of ANR debugging.
Third, some complicated ANR issues are hard to reproduce,
e.g. ANRs during MTBF test. So it's important to get as much
information as possible upon the first occurrence of this
kind of issue. The message history is very helpful to make a
good guess of the culprit. While MTBF test usually runs on
userdebug devices, so it's important to turn it on for
userdebug build.

Change-Id: I1bba71f8b751158fa392452f6dc680bcc5df2cde
Signed-off-by: Guobin Zhang <guobin.zhang@intel.com>