Update patch set 5
Patch Set 5:

Thanks for the fix.

We discussed internally and think we need to do more - an app could put a valid-looking proxy url that still doesn't work and it would pass this test.  Ideally we would verify connectivity to the proxy and verify it actually works, but that's a larger change.

Patch-set: 5