Update patch set 2
Patch Set 2:

Case #3 is not something the platform supports - you're changing things behind its back.  Couldn't you modify the EthernetStateTracker so you could set the static config and not need to use rooted-ifconfig to manually set it?  That way the framework would be in the loop and could do the right things.

Patch-set: 2