Update patch set 3
Patch Set 3: Code-Review+1

(19 comments)

One concern I have is about the DISCONNECT_CAUSE_* constants in PreciseCallState. This is the third place these constants are defined. I have another CL in review (internally in Google for now) that also needed a copy - adding call disconnect codes to the call log. That CL also makes the disc codes public and so also needed another copy. As part of that change we decided we should consolidate the duplicate definitions in a single place and so added a class android.telephony.DisconnectCause which basically just has the list of public static final int's that you have here (DISCONNECT_CAUSE_*). That replaced the Connection.DisconnectCause enum.

We should make sure this CL and mine are consistent. Could you move those constants to a new DisconnectCause class? I think it would then make sense to also move PRECISE_DISCONNECT_CAUSE_* to a PreciseDisconnectCause class. I can take care of getting rid of the existing disc code definitions as part of my CL.

Also, just as a very broad comment it is obviously not ideal to have two sets of (very similar) disconnect codes. The existing android.telephony codes are derived from the RIL codes:
 - some RIL codes are merged into one DisconnectCause
 - other codes are mapped in ways that depend on phone/uicc state
Just adding the underlying RIL code is simple and safe but adds entropy. Maybe it's the best we can do.

Patch-set: 3
Label: Code-Review=+1