Update patch set 4
Patch Set 4:

Following your instructions, created android.telephony.DisconnectCause with DISCONNECT_CAUSE_* constants and android.telephony.PreciseDisconnectCause with PRECISE_DISCONNECT_CAUSE_* constants.
Also removed constants from android.telephony.PreciseCallState.

About having two sets of disconnect cause. I agree with you that it's not convenient to have these two sets of codes. We think that a possibility is to only have the precise disconnect codes and implement methods to obtain and group the other high level/derived ones. In this way, the only set to maintain is the precise one. But I think that this is out of the scope of the current CL.

Please, report any possible inconsistency in the current changes so we can fix it ASAP.

Patch-set: 4
Label: Verified=0