Remove use of "Settings.Global.ASSISTED_GPS_ENABLED"
"Settings.Global.ASSISTED_GPS_ENABLED" is initialized only when
the assisted GPS menu is present in the Location settings menu.

This menu is not available by default.
It is usually instantiated in engineering builds to allow deactivation
of assisted GPS.
Moreover, this optional menu has been completely removed in Android 4.4.
We need to remove the use of "Settings.Global.ASSISTED_GPS_ENABLED" to
be aligned.

Change-Id: I40ab90418dd08f15e7721bbc0465c035bf4492b9
Signed-off-by: Nicolas SUET <nicolas.suet@intel.com>