Update patch set 3
Patch Set 3: -Code-Review

Thanks for the explanation. I took a look at the system properties implementation and it's not trivial to guarantee the sort of behaviour we want. 

(1) There's an asymmetry between get() and put() operations. put() reports the change to a properties socket (hosted in the init process) whereas get() reads from an mmaped file that the "server" writes to.

(2) property_set doesn't always wait for the server to write the property. It returns after a maximum of 250ms.

-1 withdrawn.

Patch-set: 3
Reviewer: Gerrit User 1014443 <1014443@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0