AArch64: Use long for pointers in graphics/shader classes
Long is used to store native pointers in graphics/Shader
classes to support 64-bit architecture.

In addition, classes (Paint and GLES20Canvas) which directly
access native pointers of shader classes have been changed.

Note that other native pointers of Paint and GLES20Canvas
classes have not been changed to long. This will be done
as part of a separate patch.

Change-Id: Ic478c65637f26133d1ca3e1c9b462960dd7b673a
Signed-off-by: Ashok Bhat <ashok.bhat@arm.com>
Signed-off-by: Marcus Oakland <marcus.oakland@arm.com>