Update patch set 1
Patch Set 1:

so an example of this problem is seen in the Camera application.  after using the viewfinder, the softkeys are left in MODE_LIGHTS_OUT.  the reason is the mMode is updated only when the transition happens through the onTransition call.  but this call will not apply when the low_ram config is set, hence in this case the transition from MODE_OPAQUE to mode MODE_LIGHTS_OUT is not carried, but the mMode has been updated as if it had happened.

feel free to suggest how this should be solved otherwise.

ps: this problem can be seen on Nexus 4 as well, granted you set the ro.config.low_ram, perhaps it will make more sense a you see it?

thanks

Patch-set: 1
Reviewer: Gerrit User 1025044 <1025044@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0