Update patch set 6
Patch Set 6:

I checked the option of attaching it on gerrit but didn't found anything.

It's possible to publish the app as a code.google.com project or GitHub like site, but it looks an overkill to me.

Anyway, we are open to meet with the requirement you might propose.

Patch-set: 6