Crash when setting AwesomeImage as wallpaper
When setting an awesome image (gif) image as wallpaper
in portrait an exception is thrown and the wallpaper
is not set.

The gif image codec does not support regional decoding,
so when cropping the image this is done by decoding the
wallpaper bitmap to memory, then cropping it to the size
needed. If the image smaller than the screen the cropping
will try to crop outside the image causing
Bitmap.createBitmap to throw an IllegalArgumentException.

This is caused by an incorrect calculation of the crop
rectangle, the height is calculated using the height of
the CropView which does not cover the whole screen (it
the screen height - navbar), later the height in portratit
mode is adjusted using defaultWallpaperSize.y, causing
the crop rectangle to be outside the height of the image.

Change-Id: I88f9e34467bd0da2ee5093e8d761664010c0f6ff