Update patch set 2
Patch Set 2:

The problem I'm trying to solve is, nowhere in AOSP is TETHER_DUN_REQUIRED set, so it defaults to 2, not set. in Kit-Kat the DUN settings were fixed so they actually work. If no provisioning is required the setting must actually be set to 0. This patch is attempt to do so in a non-persistent way.

Patch-set: 2
Reviewer: Gerrit User 1025718 <1025718@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0