Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

If this is happening, "all the time,"  then there's something else going on.  More than likely, something else is holding the wakelock while the screen is off.  This code, in itself, shouldn't be keeping the device awake or burning much power.  It should only be called once per minute while the CPU is running and keyguard is showing.  Are you seeing something different?

That being said,  as written, I'm afraid this will result in the time being stale for up to a minute after the display comes on. If not, then you're getting lucky because something else is triggering the status area to refresh.

Patch-set: 1
Label: Code-Review=-1