Update patch set 2
Patch Set 2:

It's defined by us, but it's also essentially a public API, and I'm reluctant to change it if we don't have to.  Further, while we're not sticking anything but zero/NULL in there today, I don't know that we won't want to be using a pointer once we have a display other than EGL_DEFAULT_DISPLAY.

So if we can make it all work as void* without too much trouble that would be preferred.  Leaving EGL_DEFAULT_DISPLAY as an integer constant is fine.

Patch-set: 2