Update patch set 2
Patch Set 2: Code-Review-1

(9 comments)

I think this is a bad idea, especially with the feature turned on for all non-user builds.  It's going to cause us to generate and store strings for each and every message that the main looper handles.

That's going to impact performance and jank in an awful way for developers.

Is this information really all that beneficial?  Usually I can figure out what was happening before an ANR by reverse engineering and guessing based on where the hang happened and what showed up in the log.

Patch-set: 2
Reviewer: Gerrit User 1003957 <1003957@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1