Update patch set 2
Patch Set 2: -Code-Review

(9 comments)

First, thank you Jeff for your time!
It does introduce overhead to generate and store strings, with the current implementation. Optimization can be done to reduce the overhead to the least.
Considering the incoming rate of message injection, trivial overhead may be acceptable.
This debug utility has run for 1+ year at Intel. Its improvement for debug efficiency is obvious.
First, ANR may account for about 10% of all the bugs reported during device development. It is important to improve the debugging efficiency for these bugs. With these logs, ANR analyzer can be developed to debug ANR issues automatically. With advanced debugging skills programmed into the analyzer. Text report and visual profiling graph can be produced to help identify the culprit more accurately, easier and faster.
Second, huge number of developers are involved in developing Android devices. Few of them are skilled at debugging all sorts of ANR issues. These logs and tools definitely would reduce the complexity of ANR debugging. This kind of utilities are pursued by all Android OEMs.
Third, some complicated ANR issues are hard to reproduce, e.g. ANRs during MTBF test. So it's important to get as much information as possible upon the first occurrence of this kind of issue. The message history is very helpful to  make a good guess of the culprit. While MTBF test usually runs on userdebug devices, so it's important to turn it on for userdebug build.
In the end, thank you again for reviewing my patch.
I will improve the patch according to your comments. And I will also optimize the implementation to reduce the overhead to the least. Thanks!

Patch-set: 2
Label: Code-Review=0