Update patch set 2
Patch Set 2:

Jim, agree with your concern.
We would like to concern the case that screen is off but device not suspend, e.g. music background play.
How about refresh the UI once the screen is on?
The purpose is the reduce the UI refresh when screen off to save power.

Patch-set: 2
Label: Verified=0