Update patch set 3
Patch Set 3:

As rightly noted, egl code in framework/base/ is generated from frameworks/native/opengl/tools/glgen. These tools are not equipped to generate TODO comments or {@hide} API comments. Adding support for such features will not only be tricky but also not useful in the long run as they are for  temporary fix. Any suggestions to handle this issue is welcome.

Patch-set: 3