Update patch set 3
Patch Set 3:

The deprecated functions are not temporary fixes -- removing public APIs can break apps, so they will need to be around for a while.

The easiest approach is to update the code generator to reflect the new reality (e.g. change JType.java to emit "long" instead of "int" for EGLNativeDisplayType) and then add definitions for the deprecated "int" functions by hand.

Each generated file has a header that is copied in.  The deprecated functions can be added there.  Not the preferred approach, but it'll do for now.  For EGL14 the relevant files are:

  frameworks/native/opengl/tools/glgen/stubs/egl/EGL14Header.java-if
  frameworks/native/opengl/tools/glgen/stubs/egl/EGL14cHeader.cpp

This doesn't help with the @hide annotations on the new methods.  I think you have two choices: (1) update the API definition (current.txt) at the same time you check these in; or (2) do a trivial hand-edit on the generated sources to add @hide before checking into frameworks/base, with the assumption that somebody (Narayan?) will be stripping them out and updating the API description shortly afterward.

Patch-set: 3