Support for new gTLDs
Add support for new generic Top-Level Domains (gTLDs) from ICANN.
The last time this was updated was 2011.
In 2013 and 2014 the web got and will get many more gTLDs.
Without this support many new gTLDs are not clickable in Android
and needs to be copied and pasted to reach them by the user.

Change-Id: Ic3c533883037c767e972b28eb128040648ebbd3e
Signed-off-by: Marvin Reimer <therealmarv@gmail.com>