Update patch set 1
Patch Set 1:

There are a couple of issues with PreferenceGroup. First, the docs do not specify it is thread safe. I would expect that the preferences framework is not thread safe. Second, this change is not enough as there are a few other places whether mPreferenceList is accessed without a lock being held.

Patch-set: 1