JNI:ImagerReader:Return the buffer to queue.
Return the buffer to the queue when failing to lock it with error
BAD_VALUE, with which the API 'acquireNextImage' of 'ImageReader' will
return 'null' without exception even the SurfaceImage object is created.

Then there will be no chance to return the buffer back later since
the caller can't get the new SurfaceImage object.

Change-Id: I07b64f6dbd28ea82e151e474a259a618f14bd5f7
Signed-off-by: Liangcai Li <liangcail@nvidia.com>