Update patch set 1
Patch Set 1:

> Where do you see this being used ?

Mainly by the Package Manager Service - or more specifically in the scanPackageLI() method of the PackageManagerService class (services/java/com/android/server/pm/PackageManagerService.java).

The Android runtime may need to distinguish and launch Apps using the secondary Zygote (64-bit Zygote in a mainly 32-bit system or vice versa).

In order for the Android runtime to distinguish such Apps from the others, we would need to mark these Apps on installation - i.e. in scanPackageLI().

This method can be used to decide whether a given App must be marked; by looking through the ABIs included in an .apk and checking for existence or non-existence of particular ABIs under its /lib directory.

Furthermore, the ApplicationInfo class and the "packages.xml" file could be used to cache the results of this check such that the system will not require to run this check on every launch attempt.

Patch-set: 1
Label: Verified=0