aido: Support building with updated host sysroot.
This patch fixes tools/aidl/Type.cpp to compile with
a more recent SDK host toolchain sysroot (that has not
been submitted yet).

The main issue is that the code uses ssize_t which is
defined in <sys/types.h>, but didn't include the header
directly.

Apparently, this no longer compiles when using the Ubuntu
10.04 libc6-dev headers (the current SDK toolchain is
based on Ubuntu 8.04), so perform an explicit include
to fix this.

Change-Id: Idcacf8f8d1c606b39bf8aa75ab409aa0c2a3cd5a