Update patch set 2
Patch Set 2: Code-Review+1

Hi Svetoslav/Elliott,

although fragile, the CTS test linked above shows that the digits added in this upload are digits.  If there is nothing else that needs to be done, and if you have the time, would it be possible for you to merge this?

Thanks!

/Ed

Patch-set: 2
Reviewer: Gerrit User 1004882 <1004882@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1