Update patch set 1
Patch Set 1:

Yes Brian. In some cases, although not in all scenarios, we would need to check more than one ABI. To generalize for all possibilities, I am going to rewrite the native function as follows:

The function accepts an ordered array of ABIs to search - with the first one at index 0 being the first one to compare. The return value would indicate which ABI was found - if any - returning the 0-based index of the matched entry. A value of -1 would mean no match and a value of -2 would indicate an error in opening and reading the .apk file.

Patch-set: 1