Update patch set 2
Patch Set 2:

Let's get the balling rolling with this change. I'll be uploading similar changes to deal with the other sepolicy files in the relevant libselinux code. 

For now I'm doing a simple string comparison of the selinux_version file contents. Not sure on thoughts about this file or its contents....

Patch-set: 2
Label: Verified=0