Update patch set 2
Patch Set 2:

I was thinking we could create the version file as part of external/sepolicy/Android.mk with the contents of ro.build.id or ro.build.date.utc.
This is purely an equality comparison, i.e. do not use the /data/security policy unless it was for the same build as the root policy.  Do we need anything more complicated?

Patch-set: 2