re-adding order modified to be identical to the removing order
when windows are rebuilt or re-added in some reason, child windows
of window is temparary removed in descending order. however,
child windows are re-added in ascending order. this might not be
a critical issue but not make sense at all.
this patch is to fix that by changing the order of re-adding as
descending. 

Change-Id: If3a7ee6a5b04e45aa4431f474a8128e316a357d3