Update patch set 1
Patch Set 1:

SurfaceFlinger notes:

    // if we have secure windows on this display, never allow the screen capture
    // unless the producer interface is local (i.e.: we can take a screenshot for
    // ourselves).

We allow screen shots for "recents" and rotation animations, but we don't allow them for DDMS or the screencap command.  During the rotation we're showing a screen shot of the secure layer, not the secure layer itself.

It might make more sense to fix this in SurfaceFlinger.

Patch-set: 1