NPE in GestureOverlayView
When mCurrentGesture is null, fireOnGesturePerformed is being called by
GestureOverlayView which causes an NPE and the app to force close.
This happens when you clear the overlay in the OnGestureListener using
gestureOverlay.clear(false), and adding a null check in FadeOutRunnable
for this case fixes the issue.  Refer issue #6572 for reproducing.

Signed-off-by: Chander S Pechetty <cspechetty@gmail.com>