Update patch set 1
Patch Set 1:

We are developing an application for an OEM product that requires us to pass in multiple base packages (-I flags). If the packages are not listed in ascending order based on their package id we encounter this issue and an incorrect error is reported.

Patch-set: 1
Reviewer: Gerrit User 1040628 <1040628@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0