WiFi Scan Only option does not work after leaving APM when Wifi is Off
   When ScanAlwaysAvailable being checked, wifi will still
do the scanning while wifi is off.But wifi will not resume
the scanning after leaving air plane mode when wifi is off
and ScanAlwaysAvailable being checked.
Reprudoce steps:
1. Wifi is off
2. ScanAlwaysAvailable is checked
3. Air plane mode -> on
4. Air plane mode -> off
Issue:
wifi scanning is not resumed after leaving air plane mode

So add the check for ScanAlwaysAvailable and resume the
StaDisabledWithScanState if ScanAlwaysAvailable is checked.

Signed-off-by: jdavyx <jeanx.davy@intel.com>
Signed-off-by: chenxi2x <xiaolongx.chen@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Change-Id: I2a56ec20f0daea291c0995ed619ad070c7fa80aa