Update patch set 2
Patch Set 2:

OK, I see a case in which this may happen. Looking at unbindLocked() and more specifically:

mRemoteInstance = null;
mContext.unbindService(mServiceConnection);

The code wrongly clears the remote instance before unbinding. So if the service dies right before the unbind is called, the onServiceDisconnected callback will be invoked which is written with the assumption that it should never be called if we have no handle to the remote instance.

The right change is to swap the two lines above. Can you please do that as this is the correct fix for the bug. Thanks in advance.

Patch-set: 2