PowerUI: Notify user during discharge while charger is connected
Sometimes, the current supplied by the charger will be lesser than the current
consumed by the platform. Since charging is enabled by h/w, the status would
remain as charging. So user may not be aware that, the battery is discharging
in this case. User should be notified in this case. This patch notifies the
user when the capacity drops while the charger is plugged in. It is displayed
only once as long as the charger is connected.

Change-Id: I845150739e5d67c861dc8501875702cdd05a150e
Signed-off-by: Saranya Gopal <saranya.gopal@intel.com>
Signed-off-by: Pavan Kumar S <pavan.kumar.s@intel.com>
Signed-off-by: Jenny TC <jenny.tc@intel.com>