WIP: NOT INTENDED FOR COMMIT
See https://docs.google.com/a/google.com/document/d/1O804BZZTkt7o-YuK8HFipZIRmQFSTjFabV_TMiWNnSc/edit#
step 1.

This is to allow easy comparison of the Base64 code before and after
it is moved to libcore.util. The only difference between this code and
the ones to be moved there should be the package.

Change-Id: I25eecdd871f0c026d1271df6e631a4d5942436e7