app_process: change app_process a symlink of primary version
As suggested by Brian Carlstrom, change app_process to be a symlink to the primary version
and have app_process32 as the 32-bit version.
This will make it possible to use app_process for am/pm commands when we have both app_process32
and app_process64 there and using the app_process64 as the primary version.

And as suggested by Narayan Kamath, use TARGET_PREFER_32_BIT_APPS to decide which version to use

Change-Id: Ibef545994eb95ddc82101c95df613ad1d06345e6
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>