Fix a race bug in clearGarbage of PatchCache
In clearGarbage of PatchCache, the garbage patch was returned back
to mFreeBlocks, and then removed from the cache. Since this function
is called from another thread other than the rendering thread. There
is a possibility that the cache was hit and reused by rendering
thread after it was returned to mFreeBlocks but before removing from
cache. As a result, the rendering thread gets wrong VB, UI corruption
is observed. To fix it, remove it from cache before return it back
to mFreeBlocks.

Change-Id: Idef4af737648f5d7128ac52088707357025089eb