Fix IndexOutOfBoundsException introduced by ProcessStats.
As we know, multiple packages can run in a single process, such as package 
com.android.providers.media and package com.android.providers.downloads. 
If we use package com.android.providers.media to start a process, whose 
ProcessState and CommonProcessState are same, and will be marked as active. 
If this process died, whose ProcessState and CommonProcessState are marked 
as inactive. 

After this, we use package com.android.providers.downloads to restart the 
above-mentioned process, it will use above inactive CommonProcessState to 
create a new active ProcessState for package com.android.providers.downloads, 
and to clone a new inactive ProcessState for package com.android.providers.media. 

If we restart package com.android.providers.media whose ProcessState will 
be marked as active, but CommonProcessState is still inactive. After 
resetSafely function of ProcessStats called, the CommonProcessState 
of package com.android.providers.media will be marked as dead, 
and mLongs of ProcessStats will be reset, but mDurationsTable of 
CommonProcessState is not reset, so it maybe throw IndexOutOfBoundsException 
when you use index got from mDurationsTable of CommonProcessState to access 
reseted mLongs array of ProcessStats, such as call addDuration function.

This patch will resonablely update active flag of CommonProcessState 
when call makeActive or makeInactive function of ProcessState through 
a new added mRefCnt member, which record count of active ProcessState 
created or cloned from this CommonProcessState, so the CommonProcessState 
cannot be marked as dead if it has at least one active ProcessState when 
call resetSafely function.

Change-Id: Iec7e4069663d2e2759e8fe72ae276c62f7bc0f1b
Signed-off-by: Liu, KaiX Z <kaix.z.liu@intel.com>
Signed-off-by: Guobin Zhang <guobin.zhang@intel.com>