Support an instruction-set flag for instrumentation.
Allows us to choose what instruction set a process
uses when launching it with "adb shell am instrument".

Note that we only perform very basic validation of the
instruction set. In general, there is no guarantee that
the app will support the instruction set we want to launch
it with.

bug: 14453227

Change-Id: Ifb7e89b53675080dc87941091ee5ac360f218d7f