Update patch set 1
Patch Set 1:

When the system runs into low memory situation. WindowManager tries to trim memory from HWUI cache. It will call HardwareRenderer.triMemory, and finally calls into cache.clearGarbage(). This call was not from the rendering thread. And when we debug this issue, it is much more easier to reproduce in low memory case.

Patch-set: 1