Update patch set 2
Patch Set 2:

Getting it working with "am start" is not really feasible because that command does not kill the existing process.

We could restrict it to adb shell am start -S which I believe kills the process, but that sounds inconsistent.

Would it be better to add a flag to adb install that forces a 32 or 64 bit install ?

Patch-set: 2