Update patch set 2
Patch Set 2:

Yes, I get that - but we would need the same mechanism for apps that bundle native libs as well because they have the same issues.

Also, even if this was supported with "start" there would be a boot load of restrictions because as I said earlier, start will not kill the process if it's already running (and some activity flags that start accepts are incompatible with killing processes).

Patch-set: 2