Update patch set 2
Patch Set 2:

Is there actually a leak here? I'm not terribly familiar with  Parcels, but it seems that at worst the current Metadata code is slightly inefficient because it allows the Parcel to be GCd instead of returning it to the pool for re-use (which might then later require a new Parcel to be allocated).

The change looks OK to me, but if there isn't actually a leak here, it would be good to update the commit message to reflect that.

Patch-set: 2