Update patch set 2
Patch Set 2:

Using /data/system/users/<userid> would be ideal, if you can make the file permissions work properly to only allow read-all to that file.

As pointed out elsewhere though, it's weird that we have to mark these files read-all for a user because the access is done directly by the client app rather than through an RPC to the keystore process. Does changing that involve changing the behavior of standard java libraries?

Patch-set: 2