Update patch set 1
Patch Set 1:

The system log was full of the following line:
 "BackupManagerService Backup time but one already running."

This log line is printed when BackupManagerService receives the
following intent: "android.app.backup.intent.RUN". This intent 
is sent from two different locations, backupnow(), startBackupAlarmsLocked().

BugReport indicate that the second method is the one creating the problem

 Dalvik Backtrace thread 492
 #0 com.android.server.AlarmManagerService.?(page not mapped) (AlarmManagerService.java:247)
 #1 android.app.AlarmManager.?(page not mapped) (AlarmManager.java:----)
 #2 com.android.server.BackupManagerService.?(page not mapped) (BackupManagerService.java:5333)
 #3 com.android.server.BackupManagerService.access$800 (BackupManagerService.java:136)
 #4 com.android.server.BackupManagerService$PerformBackupTask.?(page not mapped) 
 (BackupManagerService.java:2233)
 #5 com.android.server.BackupManagerService$PerformBackupTask.revertAndEndBackup
 (BackupManagerService.java:2201)
 #6 com.android.server.BackupManagerService$PerformBackupTask.operationComplete
 (BackupManagerService.java:2163)
 #7 com.android.server.BackupManagerService$BackupHandler.handleMessage (BackupManagerService.java:571)
 #8 android.os.Handler.?(page not mapped) (----:----)
 #9 android.os.Looper.?(page not mapped) (----:263)
 #10 android.os.HandlerThread.run (----:----)

What happens in the two attached logs are that at a given time, backup
starts (and fails) and then the backup is restarted and failing
indefinitely until the handset crashes.

The backup code itself is placed in Google code, but the logs give some
information. Obviously, a special scheduler class is used to schedule
the next update, OperationScheduler. This class has a trigger member
which decides when it's time for the next attempt. 
The printout we see is typically:

 WARN [ 4902.042210] (512:1048) BackupTransportService Not ready for backup request right now:
 [OperationScheduler: enabledState=false lastSuccess=2012-09-19/09:16:38
 moratoriumSet=2012-09-19/12:15:29 moratorium=1970-01-01/01:00:00 trigger=1970-01-01/01:00:00]

where we can see the trigger set to 1970-01-01/01:00:00. In the
OperationScheduler class, I can see that this translates to an
immediate retrigger of the scheduler, pain is taken to correct
times that are before now.

This, to me, means that the clock was not updated, but instead
the backup service got a zero time back.

This data is not recent, maybe the internal (Google) code has
changed and does not scedule the same way, but independent of
this the "requestBackupTime()" can be implemted extranally and
it is good practice to sanitize data from external sources.

Patch-set: 1